final fixes to tensor

+++ flood_forecast/basic/linear_regression.py
@@ -36,7 +36,8 @@ def simple_decode(model: Type[torch.nn.Module], src: torch.Tensor, max_seq_len: 
     :device used to to match function signature 
     :returns a torch.Tensor of dimension (B, L, M)
     """
-    real_target2 = real_target.float().clone()
+    real_target = real_target.float()
+    real_target2 = real_target.clone()
     ys = src[:, -1, :].unsqueeze(unsqueeze_dim)
     for i in range(0, max_seq_len):
         with torch.no_grad():
